- builds for ac, shared disabled on x86_64 as i can't find where -fPIC gets   _not_ passed and as it was never enabled for AC it's not regression either
Changed files:
    uClibc.spec -> 1.153.2.1

+++ uClibc.spec
@@ -1,3 +1,11 @@
+#
+# Conditional build:
+%bcond_without	shared		# don't build shared lib support
+#
+%ifarch %{x8664}
+%undefine	with_shared
+%endif
+#
 Summary:	C library optimized for size
 Summary(pl.UTF-8):	Biblioteka C zoptymalizowana na rozmiar
 Name:		uClibc
@@ -16,6 +24,7 @@ Patch4:		%{name}-stdio-unhide.patch
 Patch5:		%{name}-sparc.patch
 URL:		http://uclibc.org/
 BuildRequires:	binutils-gasp
+BuildRequires:	cpp
 BuildRequires:	gcc >= 5:3.0
 BuildRequires:	linux-libc-headers >= 7:2.6.20
 BuildRequires:	sed >= 4.0
@@ -26,8 +35,6 @@ BuildRoot:	%{tmpdir}/%{name}-%{version}-root-%(id -u -n)
 %define		uclibc_root	/usr/%{_target_cpu}-linux-uclibc
 # note: the 2nd '\' is needed (some shell expansions?)
 %define		TARGET_ARCH	%(echo %{_target_cpu} | sed -e 's/i.86\\|athlon\\|pentium./i386/;s/ppc/powerpc/;s/amd64\\|ia32e/x86_64/')
-
-%define		specflags	-fgnu89-inline
 
 %description
 Small libc for building embedded applications.
@@ -104,6 +111,9 @@ sed -e 's/^.*UCLIBC_HAS_IPV6.*$/UCLIBC_HAS_IPV6=y/;
 	s/^.*UCLIBC_HAS_RPC.*/UCLIBC_HAS_RPC=y\n# UCLIBC_HAS_FULL_RPC is not set\n# UCLIBC_HAS_REENTRANT_RPC is not set/;
 	s/^.*UCLIBC_HAS_SYS_SIGLIST.*$/UCLIBC_HAS_SYS_SIGLIST=y/;
 	s,^SHARED_LIB_LOADER_PREFIX=.*,SHARED_LIB_LOADER_PREFIX="$(RUNTIME_PREFIX)/lib",
+%if %{with shared}
+	s/^\(HAVE_SHARED\)=y/# \1 is not set/;
+%endif
 	s/^.*UCLIBC_HAS_PRINTF_M_SPEC.*$/UCLIBC_HAS_PRINTF_M_SPEC=y/;
 	s/^.*UCLIBC_SUSV3_LEGACY.*$/UCLIBC_SUSV3_LEGACY=y\nUCLIBC_SUSV3_LEGACY_MACROS=y/;
 %{?debug:s/^.*\<DODEBUG\>.*$/DODEBUG=y/;s/^.*SUPPORT_LD_DEBUG\>.*$/SUPPORT_LD_DEBUG=y/;s/^.*SUPPORT_LD_DEBUG_EARLY.*$/SUPPORT_LD_DEBUG_EARLY=y/;s/^.*\<DOSTRIP\>.*$/# DOSTRIP is not set/;}
@@ -135,18 +145,52 @@ install -d $RPM_BUILD_ROOT%{_bindir}
 	CC="%{__cc}" \
 	PREFIX=$RPM_BUILD_ROOT
 
+%if %{with shared}
 mv -f $RPM_BUILD_ROOT%{uclibc_root}/usr/lib/{libpthread-uclibc,libpthread}.so
 ln -sf libpthread-0.9.29.so $RPM_BUILD_ROOT%{uclibc_root}/lib/libpthread.so.0
+%endif
 
 # these links are *needed* (by stuff in bin/)
 for f in $RPM_BUILD_ROOT%{uclibc_root}/bin/*; do
-	mv -f $f $RPM_BUILD_ROOT%{_bindir}
-	ln -sf ../../bin/`basename $f` $f
+	if [ -L $f ]; then
+		l=$(readlink $f)
+		a=${l##*/}
+		d=${l%/*}
+		case "$d" in
+		%{_bindir})
+			ln -sf ${l#%{_bindir}/} $RPM_BUILD_ROOT%{_bindir}/${f##*/}
+			rm -f $f
+			;;
+		$a)
+			mv -f $f $RPM_BUILD_ROOT%{_bindir}
+			;;
+		*)
+			exit 1
+			;;
+		esac
+	else
+		a=${f#*/%{_target_cpu}-uclibc-}
+		ln -sf %{_bindir}/$(basename $f) $RPM_BUILD_ROOT%{uclibc_root}/usr/bin/$a
+		mv -f $f $RPM_BUILD_ROOT%{_bindir}
+	fi
 done
 
-for f in c++ cc g++ gcc ld; do
-	ln -sf /usr/bin/%{_target_cpu}-uclibc-$f \
-		$RPM_BUILD_ROOT%{uclibc_root}/usr/bin/$f
+for f in $RPM_BUILD_ROOT%{uclibc_root}/usr/bin/*; do
+	if [ -L $f ]; then
+		l=$(readlink $f)
+		case "${l%/*}" in
+		%{uclibc_root}/bin)
+			a=${l#*/%{_target_cpu}-uclibc-}
+			ln -sf %{_bindir}/$a $f
+			;;
+		%{_bindir})
+			:
+			;;
+		*)
+			exit 2
+			;;
+		esac
+	fi
 done
 
 rm -rf $RPM_BUILD_ROOT%{uclibc_root}/usr/include/{linux,asm*}
@@ -171,8 +215,10 @@ rm -rf $RPM_BUILD_ROOT
 %doc Changelog* DEDICATION.mjn3 MAINTAINERS README TODO
 %dir %{uclibc_root}
 %ifarch %{ix86} %{x8664} ppc sparc sparcv9
+%if %{with shared}
 %dir %{uclibc_root}/lib
 %attr(755,root,root) %{uclibc_root}/lib/*.so*
+%endif
 %endif
 
 %files devel
@@ -184,9 +230,11 @@ rm -rf $RPM_BUILD_ROOT
 %dir %{uclibc_root}/usr/bin
 %attr(755,root,root) %{uclibc_root}/usr/bin/*
 %dir %{uclibc_root}/usr/lib
+%if %{with shared}
 %{uclibc_root}/usr/lib/uclibc_nonshared.a
 %ifarch %{ix86} %{x8664} ppc sparc sparcv9
 %attr(755,root,root) %{uclibc_root}/usr/lib/*.so
+%endif
 %endif
 %{uclibc_root}/usr/include
